Make embedded string length a long for VWA
A short (2 bytes) will cause unaligned struct accesses when strings are
used as a buffer to directly store binary data.

+++ test/-ext-/string/test_capacity.rb
@@ -66,7 +66,7 @@ class Test_StringCapacity < Test::Unit::TestCase
 
   def embed_header_size
     if GC.using_rvargc?
-      2 * RbConfig::SIZEOF['void*'] + RbConfig::SIZEOF['short']
+      2 * RbConfig::SIZEOF['void*'] + RbConfig::SIZEOF['long']
     else
       2 * RbConfig::SIZEOF['void*']
     end
